backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["cessq"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
django = "^5.0.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"
cython = "^3.0.8"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
